app/simple-api/requirements.txt
fastapi
uvicorn
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
